requirements.txt
cv2
numpy
vidstab
torch
Pillow
torchvision